requirements.txt
# Core
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
aiofiles>=23.2.1
jinja2>=3.1.3

# ML
openai-whisper>=20231117
faster-whisper>=1.0.0
transformers>=4.39.0
torch>=2.2.0

# Audio
librosa>=0.10.1
soundfile>=0.12.1

# Data
kagglehub>=0.2.5
numpy>=1.26.0